moving around some files, adding new packages
