+++ .idea/workspace.xml
@@ -4,7 +4,17 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="d7b308d3-29c1-431e-9c6d-76e5d7a2b191" name="Changes" comment="made player icons teleport tiles one by one" />
+    <list default="true" id="d7b308d3-29c1-431e-9c6d-76e5d7a2b191" name="Changes" comment="made player icons teleport tiles one by one">
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/DefaultPlayer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/DefaultPlayer.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/PoorMan.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/PoorMan.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/clerk.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/clerk.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/collegeStudent.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/collegeStudent.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/inheritor.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/inheritor.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/landlord.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/landlord.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/view/GameBoard.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/GameBoard.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/view/StockMarketView.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/StockMarketView.java" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -16,7 +26,7 @@
   <component name="Git.Settings">
     <option name="RECENT_BRANCH_BY_REPOSITORY">
       <map>
-        <entry key="$PROJECT_DIR$" value="main" />
+        <entry key="$PROJECT_DIR$" value="nathan" />
       </map>
     </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
@@ -72,6 +82,7 @@
     "Application.GameView.executor": "Run",
     "Application.Main.executor": "Run",
     "Application.MonopolyGame.executor": "Run",
+    "Application.TileView.executor": "Run",
     "ModuleVcsDetector.initialDetectionPerformed": "true",
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "RunOnceActivity.git.unshallow": "true",
@@ -87,6 +98,7 @@
 }]]></component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/src/main/Resources" />
       <recent name="$PROJECT_DIR$" />
     </key>
   </component>
@@ -128,7 +140,23 @@
       <option name="project" value="LOCAL" />
       <updated>1752778410159</updated>
     </task>
-    <option name="localTasksCounter" value="3" />
+    <task id="LOCAL-00003" summary="minor bug fixes">
+      <option name="closed" value="true" />
+      <created>1753397835103</created>
+      <option name="number" value="00003" />
+      <option name="presentableId" value="LOCAL-00003" />
+      <option name="project" value="LOCAL" />
+      <updated>1753397835103</updated>
+    </task>
+    <task id="LOCAL-00004" summary="Player changes colour">
+      <option name="closed" value="true" />
+      <created>1753397991589</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1753397991589</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -148,6 +176,8 @@
   <component name="VcsManagerConfiguration">
     <MESSAGE value="added &quot;teleportation movement of players&quot;" />
     <MESSAGE value="made player icons teleport tiles one by one" />
-    <option name="LAST_COMMIT_MESSAGE" value="made player icons teleport tiles one by one" />
+    <MESSAGE value="minor bug fixes" />
+    <MESSAGE value="Player changes colour" />
+    <option name="LAST_COMMIT_MESSAGE" value="Player changes colour" />
   </component>
 </project>
