+++ src/main/entity/players/PoorMan.java
@@ -1,14 +1,15 @@
-package main.entity;
+package main.entity.players;
 
+import main.entity.Stock;
 import main.use_case.Player;
 
 import java.awt.*;
 
-public class collegeStudent extends Player {
-    private static final int STUDENT_INIT_MONEY = 1000;
-    public collegeStudent(String name, Color color) {
-        super(name, STUDENT_INIT_MONEY,color);
-        this.loadPortrait("main/Resources/Computer-nerd.jpg");
+public class PoorMan extends Player {
+    private static final int POORMAN_INIT_MONEY = 200;
+    public PoorMan(String name, Color color) {
+        super(name, POORMAN_INIT_MONEY, color);
+        this.loadPortrait("poormana.png");
     }
 
     @Override
@@ -31,39 +32,25 @@ public class collegeStudent extends Player {
         }
     }
 
-    /**
-     * @param basePrice
-     * @return
-     */
     @Override
     public float adjustStockBuyPrice(float basePrice) {
-        return basePrice * 0.90f;
+        return 0;
     }
 
-    /**
-     * @param basePrice
-     * @return
-     */
     @Override
     public float adjustStockSellPrice(float basePrice) {
-        return basePrice * 1.3f;
+        return 0;
     }
 
-    /**
-     * @param baseRent
-     * @return
-     */
     @Override
     public float adjustRent(float baseRent) {
-        return baseRent;
+        return 0;
     }
 
-    /**
-     * College Student has to pay his school tuition every turn.
-     */
     @Override
     public void applyTurnEffects() {
-        this.deductMoney(100);
-        System.out.println("ah man, the tuition goes up again!");
+        System.out.println("spare a dollar for a poor fellah.?");
+        this.deductMoney(20);
+
     }
 }
