+++ src/main/entity/players/clerk.java
@@ -1,14 +1,15 @@
-package main.entity;
+package main.entity.players;
 
+import main.entity.Stock;
 import main.use_case.Player;
 
 import java.awt.*;
 
-public class PoorMan extends Player {
-    private static final int POORMAN_INIT_MONEY = 20;
-    public PoorMan(String name, Color color) {
-        super(name, POORMAN_INIT_MONEY, color);
-        this.loadPortrait("poormana.png");
+public class clerk extends Player {
+    private static final int CLERK_INIT_MONEY = 1200;
+    public clerk(String name, Color color) {
+        super(name, CLERK_INIT_MONEY, color);
+        this.loadPortrait("main/Resources/clerk.jpg");
     }
 
     @Override
@@ -31,25 +32,40 @@ public class PoorMan extends Player {
         }
     }
 
+    /**
+     * Hello.
+     * @param basePrice
+     * @return
+     */
     @Override
     public float adjustStockBuyPrice(float basePrice) {
-        return 0;
+        return basePrice;
     }
 
+    /**
+     * @param basePrice
+     * @return
+     */
     @Override
     public float adjustStockSellPrice(float basePrice) {
-        return 0;
+        return basePrice;
     }
 
+    /**
+     * @param baseRent
+     * @return
+     */
     @Override
     public float adjustRent(float baseRent) {
-        return 0;
+        return baseRent;
     }
 
+    /**
+     * Clerk recieves his salary every turn
+     */
     @Override
     public void applyTurnEffects() {
-        System.out.println("spare a dollar for a poor fellah.?");
-        this.deductMoney(20);
-
+        this.addMoney(50);
+        System.out.println("Just another day");
     }
 }
