+++ src/main/entity/players/collegeStudent.java
@@ -1,18 +1,15 @@
-package main.entity;
+package main.entity.players;
 
+import main.entity.Stock;
 import main.use_case.Player;
 
 import java.awt.*;
 
-/**
- * A Player subclass representing a Landlord.
- * Landlords gain extra rent and can sell properties for more money.
- */
-public class landlord extends Player {
-    private static final int LANDLORD_INIT_MONEY = 800;
-    public landlord(String name, Color color) {
-        super(name, LANDLORD_INIT_MONEY, color);
-        this.loadPortrait("main/Resources/landlord.png");
+public class collegeStudent extends Player {
+    private static final int STUDENT_INIT_MONEY = 1000;
+    public collegeStudent(String name, Color color) {
+        super(name, STUDENT_INIT_MONEY,color);
+        this.loadPortrait("main/Resources/Computer-nerd.jpg");
     }
 
     @Override
@@ -41,7 +38,7 @@ public class landlord extends Player {
      */
     @Override
     public float adjustStockBuyPrice(float basePrice) {
-        return (float) (basePrice * 1.8);
+        return basePrice * 0.90f;
     }
 
     /**
@@ -50,7 +47,7 @@ public class landlord extends Player {
      */
     @Override
     public float adjustStockSellPrice(float basePrice) {
-        return (float) (basePrice * 0.8);
+        return basePrice * 1.3f;
     }
 
     /**
@@ -59,14 +56,15 @@ public class landlord extends Player {
      */
     @Override
     public float adjustRent(float baseRent) {
-        return (float) (baseRent * 1.8);
+        return baseRent;
     }
 
     /**
-     * Landlord has no TurnEffects.
+     * College Student has to pay his school tuition every turn.
      */
     @Override
     public void applyTurnEffects() {
-        System.out.println("Gimme the Rent!");
+        this.deductMoney(100);
+        System.out.println("ah man, the tuition goes up again!");
     }
 }
